models/requirements-lite.txt
lmdb
numpy
Pillow
tqdm
opencv-python
efficientnet_pytorch==0.7.1
timm==0.4.12
segmentation_models_pytorch==0.2.1
jpegio
torch==1.11.0+cu113 torchvision==0.12.0+cu113


# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124